bug(mydata)#29: 증권 보유종목 역직렬화 오류 및 대시보드 증권 자산 미반영 수정

+++ src/main/java/com/mydata/domain/stock/client/dto/response/StockHoldingResponse.java
@@ -15,9 +15,9 @@ public class StockHoldingResponse {
 
   private Long averagePrice;
 
-  private Long currentPrice;
+  private Double currentPrice;
 
-  private Long evaluationAmount;
+  private Double evaluationAmount;
 
   private Double profitRate;
 }
